refactor: The http_port struct member type chnaged from uint32_t to uint16_t the validation of valid port is more strict (in range 1 to 65535)

+++ hawkgated/src/util/hg_config.h
@@ -9,7 +9,7 @@
 typedef struct config
 {
     std::string iface_name;
-    uint32_t http_port;
+    uint16_t http_port;
     std::string portal_ip;
     uint64_t session_timeout;
     uint32_t max_clients;
